# go/model_consultation.go
/*
 * api
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Consultation struct {

	Id float32 `json:"id,omitempty"`

	UserId float32 `json:"userId,omitempty"`

	MentorId float32 `json:"mentorId,omitempty"`

	DesirableDate string `json:"desirableDate,omitempty"`

	Message string `json:"message,omitempty"`
}