# go/model_mentor.go
/*
 * api
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

// Mentor - Mentor's information
type Mentor struct {

	Id float32 `json:"id"`

	Username string `json:"username"`

	Email string `json:"email,omitempty"`

	Image string `json:"image,omitempty"`

	ClickCount float32 `json:"clickCount"`

	Destination string `json:"destination"`

	EduOrg string `json:"eduOrg"`

	Term string `json:"term"`

	Type []string `json:"type"`

	Price float32 `json:"price"`
}