# go/model_user.go
/*
 * api
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

// User - 
type User struct {

	Id float32 `json:"id"`

	Username string `json:"username"`

	Email string `json:"email"`

	SignUpDate string `json:"signUpDate"`
}